+++ src/pages/Captcha/CaptchaPage.tsx
@@ -1,0 +1,8 @@
+export default function CaptchaPage() {
+  return (
+    <main>
+      <h1>캡차</h1>
+      <p>스캔 전 사용자 확인을 위한 캡차 화면입니다.</p>
+    </main>
+  );
+}
