--- src/pages/CaptchaPage.tsx
@@ -1,8 +1,0 @@
-export default function CaptchaPage() {
-  return (
-    <main>
-      <h1>캡챠</h1>
-      <p>스캔 전 사용자 확인용 캡챠 화면입니다.</p>
-    </main>
-  );
-}
