+++ src/pages/Loading/LoadingPage.tsx
@@ -2,7 +2,7 @@ export default function LoadingPage() {
   return (
     <main>
       <h1>로딩 화면</h1>
-      <p>QR 분석 진행 상태와 단계별 메시지가 표시될 로딩 화면입니다.</p>
+      <p>QR 분석 진행 상태와 단계별 메시지를 표시하는 로딩 화면입니다.</p>
     </main>
   );
 }
