+++ src/pages/Report/ReportPage.tsx
@@ -2,7 +2,7 @@ export default function ReportPage() {
   return (
     <main>
       <h1>상세 보고서 페이지</h1>
-      <p>위험 요소, 세부 지표, 대응 가이드가 표시될 상세 보고서 페이지입니다.</p>
+      <p>위험 요소, 대응 지침, 추가 가이드를 제공하는 상세 보고서 페이지입니다.</p>
     </main>
   );
 }
