+++ src/pages/ResultCritical/ResultCriticalPage.tsx
@@ -1,0 +1,8 @@
+export default function ResultCriticalPage() {
+  return (
+    <main>
+      <h1>결과 화면 - 위험</h1>
+      <p>분석 결과가 위험으로 분류된 경우 표시되는 결과 화면입니다.</p>
+    </main>
+  );
+}
