--- src/pages/ResultCriticalPage.tsx
@@ -1,8 +1,0 @@
-export default function ResultCriticalPage() {
-  return (
-    <main>
-      <h1>결과 화면 - 심각</h1>
-      <p>분석 결과가 심각으로 분류된 경우의 결과 화면입니다.</p>
-    </main>
-  );
-}
