+++ src/pages/ResultWarning/ResultWarningPage.tsx
@@ -2,7 +2,7 @@ export default function ResultWarningPage() {
   return (
     <main>
       <h1>결과 화면 - 주의</h1>
-      <p>분석 결과가 주의로 분류된 경우의 결과 화면입니다.</p>
+      <p>분석 결과가 주의로 분류된 경우 표시되는 결과 화면입니다.</p>
     </main>
   );
 }
