+++ src/pages/ScanHistory/ScanHistoryPage.tsx
@@ -1,0 +1,8 @@
+export default function ScanHistoryPage() {
+  return (
+    <main>
+      <h1>스캔 이력 목록</h1>
+      <p>최근 스캔 내역과 상태 배지를 표시하는 스캔 이력 목록 페이지입니다.</p>
+    </main>
+  );
+}
