--- src/pages/ScanHistoryPage.tsx
@@ -1,8 +1,0 @@
-export default function ScanHistoryPage() {
-  return (
-    <main>
-      <h1>스캔 이력 목록창</h1>
-      <p>최근 스캔 내역과 상태 배지가 표시될 스캔 이력 목록창입니다.</p>
-    </main>
-  );
-}
